pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gpumcrpt"
version = "0.1.0"
description = "GPU-accelerated Monte Carlo radionuclide particle transport dosimetry (research prototype)."
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [
  { name = "GPUMCRPTDosimetry contributors" }
]
classifiers = [
  "Programming Language :: Python :: 3",
  "Programming Language :: Python :: 3 :: Only",
]

dependencies = [
  "nibabel>=5.2",
  "numpy>=1.26",
  "torch>=2.2",
  "triton>=2.1",
  "h5py>=3.10",
  "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
  "pytest>=7.4",
]

[tool.setuptools]
package-dir = {"" = "src"}
include-package-data = true

[tool.setuptools.packages.find]
where = ["src"]
include = ["gpumcrpt*"]

[tool.pytest.ini_options]
addopts = "-q"
